moved FlexibleContentUnion type definitions to Components
before we would have to define the Flexible Content Union's types next to it as to reduce accidental
including of the GraphQL based files which have a Redis dependancy which doesn't work on front end!

bugt I moved it back into there for simplicity and easier new dev uptake. When we move to webpack 4
tree shaking should fix acciedental includes in the future.

also the new version of apollo in memory cache added an ugly warning about how it cannot determine how
to properly set my cache values as it doesn't understand the difference between Union/Interface results
and therefore has no way to infer. I therefore added code which pulls the required info (fragment types)
for the cache to work properly on a apollo graphql in memory cache which has no root schema definition access.